+++ src/App.js
@@ -1,4 +1,4 @@
-import './App.css';
+import './styles/App.css';
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Header from './components/Header';
