+++ src/components/Header.js
@@ -1,4 +1,4 @@
-import './Header.css';
+import '../styles/Header.css';
 import { NavLink } from 'react-router-dom';
 
 const Header = () => (
