+++ src/components/Home.js
@@ -1,4 +1,4 @@
-import './Home.css';
+import '../styles/Home.css';
 
 const Home = () => (
   <div className="site-home">
